ffi: fix use-after-free
The `ImageView` _borrows_ from the image buffer in `DecodeBarcodeParams` --
the previous code freed it too early.

To help prevent some of these issues in the future, we'll be a bit clearer
that `DecodeBarcodeParams` _owns_ the buffer and actually free it in the
destructor. Same with `EncodeBarcodeParams` and `contents`.

I've also disabled the copy/move constructors so we don't accidentally
double free or free a nullptr.