Support Android devices that use YUV420 with plane.bytesPerRow != image.width
Some devices such as the Moto G Stylus 5G (2022) provide a CameraImage that has a YUV420 image where the Y plane bytesPerRow doesn't match the image width. In these cases zxing_cpp does not seem to interpret the image correctly and fails to recognize barcodes. This workaround converts the y plane to a tightly packed Uint8List for processing instead in these cases.